Corrigido recycler view
