+++ app/build.gradle
@@ -37,6 +37,10 @@ android {
     buildFeatures{
         viewBinding true
     }
+
+    dataBinding {
+        enabled = true
+    }
 }
 
 dependencies {
